+++ Arbiter.Interop/Window/NativeWindow.cs
@@ -1,0 +1,9 @@
+﻿namespace Arbiter.Interop.Window;
+
+public struct NativeWindow
+{
+    public IntPtr Handle { get; set; }
+    public int ProcessId { get; set; }
+    public required string ClassName { get; set; }
+    public required string Title { get; set; }
+}